LayoutTests/fast/page-color-sampling/color-sampling-includes-subframes.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            font-family: system-ui;
        }

        header {
            text-align: center;
            position: fixed;
            top: 0px;
            left: 0;
            font-size: 32px;
            width: 100%;
            height: 250px;
        }

        iframe {
            width: 100%;
            height: 100px;
        }

        iframe.bottom {
            position: fixed;
            bottom: 0;
            left: 0;
            right: 0;
        }

        .tall {
            width: 10px;
            height: 2000px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(100, 0, 100, 0);

        let topFrame = document.querySelector("iframe.top");
        await UIHelper.callFunctionAndWaitForEvent(() => {
            topFrame.srcdoc = `
                <body style='margin: 0; width: 100%; height: 100%; background: rgb(255, 100, 0);'>
                    <p>Hello world</p>
                </body>`;
        }, topFrame, "load");

        let bottomFrame = document.querySelector("iframe.bottom");
        await UIHelper.callFunctionAndWaitForEvent(() => {
            bottomFrame.srcdoc = `<!DOCTYPE html>
            <html>
                <head>
                    <style>
                    body, html {
                        margin: 0;
                        width: 100%;
                        height: 100%;
                        background: rgb(0, 100, 250);
                    }
                    </style>
                </head>
                <body></body>
            </html>`
        }, bottomFrame, "load");

        testPassed("Loaded iframes");

        await UIHelper.ensurePresentationUpdate();
        color = await UIHelper.fixedContainerEdgeColors();

        shouldBeEqualToString("color.top", "rgb(255, 100, 0)");
        shouldBeEqualToString("color.bottom", "rgb(0, 100, 250)");
        finishJSTest();
    });
    </script>
</head>
<body>
<header><iframe class="top" frameborder="0"></iframe></header>
<div class="tall"></div>
<iframe class="bottom" frameborder="0"></iframe>
</body>
</html>
